backend/requirements.txt
asgiref==3.5.2
certifi==2022.9.24
cffi==1.15.1
charset-normalizer==2.1.1
cryptography==38.0.4
django-case-insensitive-field==1.0.7
django-cors-headers==3.13.0
django-phonenumber-field==7.0.0
django-phonenumbers==1.0.1
djangorestframework==3.14.0
idna==3.4
phonenumbers==8.13.0
pycparser==2.21
PyMySQL==1.0.2
pytz==2022.5
requests==2.28.1
sqlparse==0.4.3
urllib3==1.26.12
django-environ==0.9.0
mysqlclient==2.0.3